program: code sectioned & float input

+++ cmd-brightness/Program.cs
@@ -10,6 +10,8 @@ namespace orez.obrightness {
 		/// </summary>
 		private static string WmiNs = "root\\WMI";
 
+
+		// static methods
 		/// <summary>
 		/// You have to dream before your dreams can come true.
 		/// : A. P. J. Abdul Kalam
@@ -34,11 +36,11 @@ namespace orez.obrightness {
 		/// </summary>
 		/// <param name="args">brightness.</param>
 		private static void Set(string[] args) {
-			byte v = 0;
-			byte.TryParse(args[0], out v);
+			double v = 0;
+			double.TryParse(args[0], out v);
 			ManagementObject o = Obj(WmiNs, "WmiMonitorBrightnessMethods");
 			if (o == null) Console.Error.WriteLine("err: WMI interfacing failed");
-			else o.InvokeMethod("WmiSetBrightness", new object[] { uint.MaxValue, v });
+			else o.InvokeMethod("WmiSetBrightness", new object[] { uint.MaxValue, (byte)v });
 		}
 
 		/// <summary>
